refactor(analysis): delete analysis/poly and move affine up a level
`extract` had one caller, the schedule pipeline's program view, and
`TileGraph` / `TileUnit` had no reader outside it. With that layer gone the
subpackage has no consumer, so all of `model`, `extract`, `access`, and
`errors` are deleted along with the public surface re-exporting them.

`affine.py` is the exception: `analysis/scope.py` and `analysis/footprint.py`
are on the real analyze path and read it. It moves to `analysis/affine.py`
unchanged, and its two importers follow. Ten symbols leave
`analysis.__all__`; `analyze()` is otherwise untouched.

The analysis spec loses § Polyhedral model, which leaves the metrics and the
composed operation as §1 and §2. That renumber changes the addresses
`tilefoundry spec analysis` resolves, so the `analyze --help` pointers and the
five specs that linked into the old numbering move with it.
