+++ docs/spec/types.md
@@ -713,4 +713,4 @@ TensorType.umat_tensor(shape, dtype)   # ranked: a shape vector
     equal to `umat_tensor`.
   - An operand carrying `UMAT` MUST NOT be charged to a memory level by the
     residency of its own type alone; what charges it is where it is consumed
-    ([analysis §2.2.1](./analysis.md#221-compute-cost)).
+    ([analysis §1.2.1](./analysis.md#121-compute-cost)).
