+++ src/tilefoundry/analysis/affine.py
