--- src/tilefoundry/analysis/poly/__init__.py
@@ -1,26 +1,0 @@
-"""Public polyhedral analysis API."""
-
-from .access import (
-    AccessFootprint,
-    AxisExtent,
-    access_footprints,
-    carried_distances,
-    statement_time_dims,
-    time_extents,
-)
-from .errors import ExtractError
-from .extract import extract
-from .model import TileGraph, TileUnit
-
-__all__ = [
-    "AccessFootprint",
-    "AxisExtent",
-    "ExtractError",
-    "TileGraph",
-    "TileUnit",
-    "access_footprints",
-    "carried_distances",
-    "extract",
-    "statement_time_dims",
-    "time_extents",
-]
